Fix small typo in plotting code
This error did not change the inference, only prevented the code from running. I made a last minute change before submission to the ests object to make, e.g., "1-3" be stored as "1 & 3" to make the csv output easier to work with (it would convert 1-3 to a date). But, code later relied on the "1-3" syntax and I didn't fix this at the time.